leetcode-fast-api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.28.2
jinja2>=3.1.2
itsdangerous>=2.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
